Add a new uxas and a working theory of sequences.
This new work depends on kind2's lustre extensions and is not compatible
with jkind.

The theory of sequences is incomplete and imprecise but consistent. The
theory is strong enough to prove interesting consistency properties
about itself and strong enough to prove properties for the waypoint
manager using a definition of `next_waypoint` that uses iteration over
the sequence indices.

The new uxas3.lus is substantially the same as uxas2, but uses kind2
extensions to clean things up (e.g., introduce free constants) and uses
the theory of sequences to represent routes.