Compositional proof over WM and ARV using new mdl.
This is not complete (there are missing properties) and not perfect (the
properties proved for the waypoint manager aren't as strong as they need
to be). However, this version of the model is consistent and does prove.
It illustrates some interesting behaviors and features, including the
need for a lemma to prove the compositional property of interest in
uxas3.lus

Work on a stronger version of the waypoint-manager is ongoing, but
requires that message IDs uniquely identify messages, which is currently
true for input messages (under model assumptions) but not for messages
in general. Without the generalized restriction, selection of instances
for messages in existentially quantified properties fails to resolve the
expected instance, leading to a meaningless counterexample.