Improved WM and ARV and initial compositional prop
All free constants are now messages rather than message elements, which
strengthens our properties: it's stronger to say that a whole message
was seen in history rather than to say that specific elements of a
message were seen in history.

To do this, we have to rely on an axiom of IDs: equality of message IDs
implies equality of messages. This is something that must be an axiom in
our current model. We could actually enforce this with our model of
input, but that's not enough to constrain existential quantification. So
we insert an assertion that whenever a free message's ID was seen on
input, the free message itself must have been on input. This is strong
enough to prove our properties.

The result of this enhancement is much cleaner and easier-to-believe
properties, especially in the Waypoint Manager. Indeed, the new
requirement for the Waypoint Manager is very clear and very clearly what
we want.

We are also able to prove a compositional property in uxas3 that a
mission command transitively references a unique automation request with
a very clean property - much more so than before.

By moving the free constant out of the Waypoint Manager node, we are
able to use it in the guarantee for the Waypoint Manager, which allows
us to restate the requirement as part of the contract. Unfortunately,
compositional analysis is not doing what I expect; I clearly need to
look at this more carefully.